refactor(Terminal): extract TerminalPane into focused subcomponents
- Create TerminalHeader component (350 lines) for header UI rendering
- Create useTerminalLogic hook (186 lines) for state and event handlers
- Refactor TerminalPane (660→310 lines) to use extracted components
- Fix Escape key cancellation in title editing (properly reverts changes)
- Fix double-click on input text triggering unwanted maximize
- Add Cancel button to context injection progress UI
- Remove unused setPerformanceMode in TerminalSettingsTab
