+++ src/components/Terminal/TerminalPane.tsx
@@ -1,44 +1,14 @@
-import React, { useState, useCallback, useEffect, useRef } from "react";
+import React, { useCallback, useEffect, useRef } from "react";
 import { useShallow } from "zustand/react/shallow";
-import {
-  Terminal,
-  Command,
-  X,
-  Maximize2,
-  Minimize2,
-  Copy,
-  ArrowDownToLine,
-  Loader2,
-  Settings2,
-} from "lucide-react";
-import {
-  DropdownMenu,
-  DropdownMenuContent,
-  DropdownMenuCheckboxItem,
-  DropdownMenuTrigger,
-  DropdownMenuLabel,
-} from "@/components/ui/dropdown-menu";
-import {
-  ClaudeIcon,
-  GeminiIcon,
-  CodexIcon,
-  NpmIcon,
-  YarnIcon,
-  PnpmIcon,
-  BunIcon,
-} from "@/components/icons";
 import type { TerminalType } from "@/types";
 import { cn } from "@/lib/utils";
-import { getBrandColorHex } from "@/lib/colorUtils";
 import { XtermAdapter } from "./XtermAdapter";
 import { ArtifactOverlay } from "./ArtifactOverlay";
-import { StateBadge } from "./StateBadge";
-import { ActivityBadge } from "./ActivityBadge";
+import { TerminalHeader } from "./TerminalHeader";
 import { ErrorBanner } from "../Errors/ErrorBanner";
-import { useErrorStore, useTerminalStore, getTerminalRefreshTier, type RetryAction } from "@/store";
-import { useContextInjection } from "@/hooks/useContextInjection";
+import { useErrorStore, useTerminalStore, getTerminalRefreshTier } from "@/store";
+import { useTerminalLogic } from "@/hooks/useTerminalLogic";
 import type { AgentState } from "@/types";
-import { errorsClient } from "@/clients";
 import { terminalInstanceService } from "@/services/TerminalInstanceService";
 
 export type { TerminalType };
@@ -70,42 +40,6 @@ export interface TerminalPaneProps {
   location?: "grid" | "dock";
 }
 
-interface TerminalIconProps {
-  className?: string;
-  brandColor?: string;
-}
-
-function getTerminalIcon(type: TerminalType, props: TerminalIconProps) {
-  const finalProps = {
-    className: cn("w-3.5 h-3.5", props.className),
-    "aria-hidden": "true" as const,
-  };
-
-  const customIconProps = { ...finalProps, brandColor: props.brandColor };
-
-  switch (type) {
-    case "claude":
-      return <ClaudeIcon {...customIconProps} />;
-    case "gemini":
-      return <GeminiIcon {...customIconProps} />;
-    case "codex":
-      return <CodexIcon {...customIconProps} />;
-    case "npm":
-      return <NpmIcon {...finalProps} />;
-    case "yarn":
-      return <YarnIcon {...finalProps} />;
-    case "pnpm":
-      return <PnpmIcon {...finalProps} />;
-    case "bun":
-      return <BunIcon {...finalProps} />;
-    case "custom":
-      return <Command {...finalProps} />;
-    case "shell":
-    default:
-      return <Terminal {...finalProps} />;
-  }
-}
-
 function TerminalPaneComponent({
   id,
   title,
@@ -119,22 +53,14 @@ function TerminalPaneComponent({
   onFocus,
   onClose,
   onInjectContext,
-  onCancelInjection: _onCancelInjection,
+  onCancelInjection,
   onToggleMaximize,
   onTitleChange,
   onMinimize,
   onRestore,
   location = "grid",
 }: TerminalPaneProps) {
-  const [isExited, setIsExited] = useState(false);
-  const [exitCode, setExitCode] = useState<number | null>(null);
-  const [isEditingTitle, setIsEditingTitle] = useState(false);
-  const [editingValue, setEditingValue] = useState(title);
-  const titleInputRef = useRef<HTMLInputElement>(null);
   const containerRef = useRef<HTMLDivElement>(null);
-
-  // Subscribe to injection progress for this terminal only
-  const { inject, isInjecting, progress: injectionProgress } = useContextInjection(id);
 
   const updateVisibility = useTerminalStore((state) => state.updateVisibility);
   const getTerminal = useTerminalStore((state) => state.getTerminal);
@@ -151,50 +77,29 @@ function TerminalPaneComponent({
   const dismissError = useErrorStore((state) => state.dismissError);
   const removeError = useErrorStore((state) => state.removeError);
 
-  const handleErrorRetry = useCallback(
-    async (errorId: string, action: RetryAction, args?: Record<string, unknown>) => {
-      try {
-        if (action === "injectContext") {
-          const worktreeIdArg = args?.worktreeId as string | undefined;
-          const terminalIdArg = args?.terminalId as string | undefined;
-          const selectedPaths = args?.selectedPaths as string[] | undefined;
+  const {
+    isEditingTitle,
+    editingValue,
+    titleInputRef,
+    setEditingValue,
+    handleTitleDoubleClick,
+    handleTitleKeyDown,
+    handleTitleInputKeyDown,
+    handleTitleSave,
+    isExited,
+    exitCode,
+    handleExit,
+    handleErrorRetry,
+    isInjecting,
+    injectionProgress,
+  } = useTerminalLogic({
+    id,
+    title,
+    onTitleChange,
+    removeError,
+  });
 
-          if (!worktreeIdArg || !terminalIdArg) {
-            console.error("Missing worktreeId or terminalId for injectContext retry");
-            return;
-          }
-
-          await inject(worktreeIdArg, terminalIdArg, selectedPaths);
-          removeError(errorId);
-        } else {
-          await errorsClient.retry(errorId, action, args);
-          removeError(errorId);
-        }
-      } catch (error) {
-        console.error("Error retry failed:", error);
-      }
-    },
-    [inject, removeError]
-  );
-
-  useEffect(() => {
-    setIsExited(false);
-    setExitCode(null);
-  }, [id]);
-
-  useEffect(() => {
-    if (!isEditingTitle) {
-      setEditingValue(title);
-    }
-  }, [title, isEditingTitle]);
-
-  useEffect(() => {
-    if (isEditingTitle && titleInputRef.current) {
-      titleInputRef.current.focus();
-      titleInputRef.current.select();
-    }
-  }, [isEditingTitle]);
-
+  // Visibility observation
   useEffect(() => {
     if (!containerRef.current) return;
 
@@ -214,58 +119,6 @@ function TerminalPaneComponent({
       updateVisibility(id, false);
     };
   }, [id, updateVisibility]);
-
-  const handleTitleDoubleClick = useCallback(
-    (e: React.MouseEvent) => {
-      e.stopPropagation();
-      if (onTitleChange) {
-        setIsEditingTitle(true);
-      }
-    },
-    [onTitleChange]
-  );
-
-  const handleTitleKeyDown = useCallback(
-    (e: React.KeyboardEvent) => {
-      if (onTitleChange && (e.key === "Enter" || e.key === "F2")) {
-        e.preventDefault();
-        e.stopPropagation();
-        setIsEditingTitle(true);
-      }
-    },
-    [onTitleChange]
-  );
-
-  const handleTitleSave = useCallback(() => {
-    if (!isEditingTitle) return;
-    setIsEditingTitle(false);
-    if (onTitleChange) {
-      onTitleChange(editingValue);
-    }
-  }, [isEditingTitle, editingValue, onTitleChange]);
-
-  const handleTitleCancel = useCallback(() => {
-    setIsEditingTitle(false);
-    setEditingValue(title);
-  }, [title]);
-
-  const handleTitleInputKeyDown = useCallback(
-    (e: React.KeyboardEvent<HTMLInputElement>) => {
-      if (e.key === "Enter") {
-        e.preventDefault();
-        handleTitleSave();
-      } else if (e.key === "Escape") {
-        e.preventDefault();
-        handleTitleCancel();
-      }
-    },
-    [handleTitleSave, handleTitleCancel]
-  );
-
-  const handleExit = useCallback((code: number) => {
-    setIsExited(true);
-    setExitCode(code);
-  }, []);
 
   const handleReady = useCallback(() => {}, []);
 
@@ -293,7 +146,6 @@ function TerminalPaneComponent({
     return getTerminalRefreshTier(terminal, isFocused);
   }, [id, isFocused, getTerminal]);
 
-  // Handler for container clicks - calls focus and boosts refresh rate
   const handleClick = useCallback(() => {
     onFocus();
     terminalInstanceService.boostRefreshRate(id);
@@ -307,16 +159,11 @@ function TerminalPaneComponent({
       className={cn(
         "flex flex-col h-full overflow-hidden transition-all duration-200 group",
         "bg-[var(--color-surface)]",
-
-        // Only apply border, rounding, and shadow if NOT in the dock (prevents double-border glitch)
         location !== "dock" && "rounded border shadow-md",
-
-        // Apply focus styles only if NOT docked
         location !== "dock" &&
           (isFocused
             ? "terminal-focused border-[color-mix(in_oklab,var(--color-canopy-border)_100%,white_20%)]"
             : "border-canopy-border hover:border-[color-mix(in_oklab,var(--color-canopy-border)_100%,white_10%)]"),
-
         isExited && "opacity-75 grayscale"
       )}
       onClick={handleClick}
@@ -347,215 +194,61 @@ function TerminalPaneComponent({
         }
       })()}
     >
-      {/* Header - Uniform background for all terminal types */}
-      <div
-        className={cn(
-          "flex items-center justify-between px-3 h-8 shrink-0 font-mono text-xs transition-colors relative overflow-hidden",
-          // Base background - uniform for all types
-          isFocused ? "bg-[var(--color-surface-highlight)]" : "bg-[var(--color-surface)]"
-        )}
-        onDoubleClick={onToggleMaximize}
-      >
-        <div className="flex items-center gap-2 min-w-0">
-          <span className="shrink-0 flex items-center justify-center w-3.5 h-3.5 text-canopy-text">
-            {isWorking ? (
-              <Loader2
-                className="w-3.5 h-3.5 animate-spin"
-                style={{ color: getBrandColorHex(type) }}
-                aria-hidden="true"
-              />
-            ) : (
-              getTerminalIcon(type, { brandColor: getBrandColorHex(type) })
-            )}
-          </span>
-
-          {isEditingTitle ? (
-            <input
-              ref={titleInputRef}
-              type="text"
-              value={editingValue}
-              onChange={(e) => setEditingValue(e.target.value)}
-              onKeyDown={handleTitleInputKeyDown}
-              onBlur={handleTitleSave}
-              className="text-sm font-medium bg-canopy-bg/60 border border-canopy-accent/50 px-1 h-5 min-w-32 outline-none text-canopy-text select-text"
-              aria-label={type === "shell" ? "Edit shell title" : "Edit agent title"}
-            />
-          ) : (
-            <span
-              className={cn(
-                isFocused ? "text-canopy-text" : "text-canopy-text/70",
-                "font-medium truncate select-none",
-                onTitleChange && "cursor-text hover:text-canopy-text"
-              )}
-              onDoubleClick={handleTitleDoubleClick}
-              onKeyDown={handleTitleKeyDown}
-              tabIndex={onTitleChange ? 0 : undefined}
-              role={onTitleChange ? "button" : undefined}
-              title={onTitleChange ? `${title} — Double-click or press Enter to edit` : title}
-              aria-label={
-                onTitleChange
-                  ? type === "shell"
-                    ? `Shell title: ${title}. Press Enter or F2 to edit`
-                    : `Agent title: ${title}. Press Enter or F2 to edit`
-                  : undefined
-              }
-            >
-              {title}
-            </span>
-          )}
-
-          {isExited && (
-            <span
-              className="text-xs font-mono text-[var(--color-status-error)] ml-1"
-              role="status"
-              aria-live="polite"
-            >
-              [exit {exitCode}]
-            </span>
-          )}
-
-          {/* Show state badge for completed/failed states (working state uses spinner instead) */}
-          {agentState &&
-            agentState !== "idle" &&
-            agentState !== "waiting" &&
-            agentState !== "working" && <StateBadge state={agentState} className="ml-2" />}
-
-          {/* Show activity headline when available and not in terminal state */}
-          {activity &&
-            activity.headline &&
-            agentState !== "failed" &&
-            agentState !== "completed" && (
-              <ActivityBadge
-                headline={activity.headline}
-                status={activity.status}
-                type={activity.type}
-                className="ml-2"
-              />
-            )}
-
-          {queueCount > 0 && (
-            <div
-              className="text-xs font-mono bg-canopy-accent/15 text-canopy-text px-1.5 py-0.5 rounded ml-1"
-              role="status"
-              aria-live="polite"
-              title={`${queueCount} command${queueCount > 1 ? "s" : ""} queued`}
-            >
-              {queueCount} queued
-            </div>
-          )}
-        </div>
-
-        <div className="flex items-center gap-1 opacity-0 group-hover:opacity-100 group-focus-within:opacity-100 transition-opacity">
-          <DropdownMenu>
-            <DropdownMenuTrigger asChild>
-              <button
-                className="p-1.5 hover:bg-canopy-text/10 focus-visible:bg-canopy-text/10 focus-visible:outline focus-visible:outline-2 focus-visible:outline-canopy-accent text-canopy-text/60 hover:text-canopy-text transition-colors rounded"
-                onClick={(e) => e.stopPropagation()}
-                title="Terminal Settings"
-                aria-label="Terminal settings"
-              >
-                <Settings2 className="w-3 h-3" aria-hidden="true" />
-              </button>
-            </DropdownMenuTrigger>
-            <DropdownMenuContent align="end" className="w-56">
-              <DropdownMenuLabel>Terminal Settings</DropdownMenuLabel>
-              <DropdownMenuCheckboxItem
-                checked={terminal?.settings?.autoRestart ?? false}
-                onCheckedChange={(checked) =>
-                  updateTerminalSettings(id, { autoRestart: checked === true })
-                }
-              >
-                Auto-restart on open
-              </DropdownMenuCheckboxItem>
-            </DropdownMenuContent>
-          </DropdownMenu>
-          {worktreeId && onInjectContext && (
-            <button
-              onClick={(e) => {
-                e.stopPropagation();
-                onInjectContext();
-              }}
-              className={cn(
-                "p-1.5 hover:bg-canopy-text/10 focus-visible:bg-canopy-text/10 focus-visible:outline focus-visible:outline-2 focus-visible:outline-canopy-accent text-canopy-text/60 hover:text-[var(--color-state-working)] transition-colors",
-                isInjecting && "opacity-50 cursor-not-allowed"
-              )}
-              title="Inject Context (Ctrl+Shift+I)"
-              aria-label="Inject worktree context"
-              disabled={isExited || isInjecting}
-            >
-              <Copy className="w-3 h-3" aria-hidden="true" />
-            </button>
-          )}
-          {onMinimize && !isMaximized && (
-            <button
-              onClick={(e) => {
-                e.stopPropagation();
-                onMinimize();
-              }}
-              className="p-1.5 hover:bg-canopy-text/10 focus-visible:bg-canopy-text/10 focus-visible:outline focus-visible:outline-2 focus-visible:outline-canopy-accent text-canopy-text/60 hover:text-canopy-text transition-colors"
-              title={location === "dock" ? "Minimize" : "Minimize to dock"}
-              aria-label={location === "dock" ? "Minimize" : "Minimize to dock"}
-            >
-              <ArrowDownToLine className="w-3 h-3" aria-hidden="true" />
-            </button>
-          )}
-          {location === "dock" && onRestore && (
-            <button
-              onClick={(e) => {
-                e.stopPropagation();
-                onRestore();
-              }}
-              className="p-1.5 hover:bg-canopy-text/10 focus-visible:bg-canopy-text/10 focus-visible:outline focus-visible:outline-2 focus-visible:outline-canopy-accent text-canopy-text/60 hover:text-canopy-text transition-colors"
-              title="Restore to grid"
-              aria-label="Restore to grid"
-            >
-              <Maximize2 className="w-3 h-3" aria-hidden="true" />
-            </button>
-          )}
-          {onToggleMaximize && (
-            <button
-              onClick={(e) => {
-                e.stopPropagation();
-                onFocus();
-                onToggleMaximize();
-              }}
-              className="p-1.5 hover:bg-canopy-text/10 focus-visible:bg-canopy-text/10 focus-visible:outline focus-visible:outline-2 focus-visible:outline-canopy-accent text-canopy-text/60 hover:text-canopy-text transition-colors"
-              title={isMaximized ? "Restore (Ctrl+Shift+F)" : "Maximize (Ctrl+Shift+F)"}
-              aria-label={isMaximized ? "Restore" : "Maximize"}
-            >
-              {isMaximized ? (
-                <Minimize2 className="w-3 h-3" aria-hidden="true" />
-              ) : (
-                <Maximize2 className="w-3 h-3" aria-hidden="true" />
-              )}
-            </button>
-          )}
-          <button
-            onClick={(e) => {
-              e.stopPropagation();
-              onClose(e.altKey);
-            }}
-            onKeyDown={(e) => {
-              if ((e.key === "Enter" || e.key === " ") && e.altKey) {
-                e.preventDefault();
-                e.stopPropagation();
-                onClose(true);
-              }
-            }}
-            className="p-1.5 hover:bg-[color-mix(in_oklab,var(--color-status-error)_15%,transparent)] focus-visible:bg-[color-mix(in_oklab,var(--color-status-error)_15%,transparent)] focus-visible:outline focus-visible:outline-2 focus-visible:outline-[var(--color-status-error)] text-canopy-text/60 hover:text-[var(--color-status-error)] transition-colors"
-            title="Close Session (Alt+Click to force close)"
-            aria-label="Close session. Hold Alt and click to force close without recovery."
-          >
-            <X className="w-3 h-3" aria-hidden="true" />
-          </button>
-        </div>
-      </div>
+      <TerminalHeader
+        id={id}
+        title={title}
+        type={type}
+        worktreeId={worktreeId}
+        isFocused={isFocused}
+        isExited={isExited}
+        exitCode={exitCode}
+        isWorking={isWorking}
+        agentState={agentState}
+        activity={activity}
+        queueCount={queueCount}
+        terminal={terminal}
+        isEditingTitle={isEditingTitle}
+        editingValue={editingValue}
+        titleInputRef={titleInputRef}
+        onEditingValueChange={setEditingValue}
+        onTitleDoubleClick={handleTitleDoubleClick}
+        onTitleKeyDown={handleTitleKeyDown}
+        onTitleInputKeyDown={handleTitleInputKeyDown}
+        onTitleSave={handleTitleSave}
+        onClose={onClose}
+        onFocus={onFocus}
+        onInjectContext={onInjectContext}
+        onToggleMaximize={onToggleMaximize}
+        onTitleChange={onTitleChange}
+        onMinimize={onMinimize}
+        onRestore={onRestore}
+        onUpdateSettings={(updates) => updateTerminalSettings(id, updates)}
+        isMaximized={isMaximized}
+        isInjecting={isInjecting}
+        location={location}
+      />
 
       {isInjecting && injectionProgress && (
         <div className="p-2 bg-canopy-sidebar border-t border-canopy-border shrink-0">
           <div className="flex items-center justify-between text-xs text-canopy-text/60 mb-1">
             <span>Injecting Context</span>
-            <span>{Math.min(100, Math.max(0, Math.round(injectionProgress.progress * 100)))}%</span>
+            <div className="flex items-center gap-2">
+              {onCancelInjection && (
+                <button
+                  type="button"
+                  onClick={(e) => {
+                    e.stopPropagation();
+                    onCancelInjection();
+                  }}
+                  className="px-2 py-0.5 rounded border border-canopy-border hover:border-canopy-accent text-canopy-text/60 hover:text-canopy-text transition-colors"
+                >
+                  Cancel
+                </button>
+              )}
+              <span>
+                {Math.min(100, Math.max(0, Math.round(injectionProgress.progress * 100)))}%
+              </span>
+            </div>
           </div>
 
           <div className="w-full h-2 bg-canopy-border rounded-full overflow-hidden mb-1">
